requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyYAML==6.0.1
requests>=2.31.0
dashscope
certifi>=2023.7.22
urllib3>=2.0.0
tqdm>=4.66.0